+++ primefaces/src/main/java/org/primefaces/model/JpaLazyDataModel.java
@@ -45,6 +45,7 @@ import javax.persistence.TypedQuery;
 import javax.persistence.criteria.*;
 
 import org.primefaces.util.BeanUtils;
+import org.primefaces.util.Constants;
 import org.primefaces.util.LangUtils;
 import org.primefaces.util.Lazy;
 import org.primefaces.util.SerializableSupplier;
@@ -62,6 +63,7 @@ public class JpaLazyDataModel<T> extends LazyDataModel<T> implements Serializabl
     protected SerializableSupplier<EntityManager> entityManager;
     protected String rowKeyField;
     protected boolean caseSensitive = true;
+    protected boolean wildcardSupport = false;
 
     private transient Lazy<Method> rowKeyGetter;
 
@@ -216,7 +218,13 @@ public class JpaLazyDataModel<T> extends LazyDataModel<T> implements Serializabl
             case NOT_CONTAINS:
                 return cb.notLike(fieldExpressionAsString.get(), "%" + getStringFilterValue(filterValue) + "%");
             case EXACT:
-                return cb.equal(fieldExpressionAsString.get(), getStringFilterValue(filterValue));
+                String exactValue = getStringFilterValue(filterValue);
+                if (wildcardSupport && (exactValue.contains("%") || exactValue.contains("_"))) {
+                    return cb.like(fieldExpressionAsString.get(), exactValue);
+                }
+                else {
+                    return cb.equal(fieldExpressionAsString.get(), exactValue);
+                }
             case EQUALS:
                 return cb.equal(fieldExpression, filterValue);
             case NOT_EXACT:
@@ -253,8 +261,14 @@ public class JpaLazyDataModel<T> extends LazyDataModel<T> implements Serializabl
         return null;
     }
 
-    protected Object getStringFilterValue(Object filterValue) {
-        return caseSensitive ? filterValue : filterValue.toString().toUpperCase(Locale.getDefault());
+    protected String getStringFilterValue(Object filterValue) {
+        String value = Objects.toString(filterValue, Constants.EMPTY_STRING);
+        value = caseSensitive ? value : value.toUpperCase(Locale.getDefault());
+        if (wildcardSupport) {
+            value = value.replace("*", "%");
+            value = value.replace("?", "_");
+        }
+        return value;
     }
 
     protected void applySort(CriteriaBuilder cb,
@@ -431,6 +445,14 @@ public class JpaLazyDataModel<T> extends LazyDataModel<T> implements Serializabl
         this.caseSensitive = caseSensitive;
     }
 
+    public boolean isWildcardSupport() {
+        return wildcardSupport;
+    }
+
+    public void setWildcardSupport(boolean wildcardSupport) {
+        this.wildcardSupport = wildcardSupport;
+    }
+
     public static <T> Builder<T> builder(Class<T> entityClass, SerializableSupplier<EntityManager> entityManager) {
         return new Builder<>(entityClass, entityManager);
     }
@@ -441,6 +463,7 @@ public class JpaLazyDataModel<T> extends LazyDataModel<T> implements Serializabl
         private String rowKeyField;
         private Converter<T> rowKeyConverter;
         private boolean caseSensitive = true;
+        private boolean wildcardSupport = false;
 
         public Builder(Class<T> entityClass, SerializableSupplier<EntityManager> entityManager) {
             this.entityClass = entityClass;
@@ -462,6 +485,11 @@ public class JpaLazyDataModel<T> extends LazyDataModel<T> implements Serializabl
             return this;
         }
 
+        public Builder<T> wildcardSupport(boolean wildcardSupport) {
+            this.wildcardSupport = wildcardSupport;
+            return this;
+        }
+
         public JpaLazyDataModel<T> build() {
             JpaLazyDataModel<T> model = new JpaLazyDataModel<>(entityClass, entityManager);
             if (rowKeyField != null) {
@@ -471,6 +499,7 @@ public class JpaLazyDataModel<T> extends LazyDataModel<T> implements Serializabl
                 model.setRowKeyConverter(rowKeyConverter);
             }
             model.caseSensitive = caseSensitive;
+            model.wildcardSupport = wildcardSupport;
             return model;
         }
     }